pyproject.toml
[project]
name = "ocarina-cards-3d"
version = "0.1.0"
description = "A parametric 3D printable Ocarina learning card generator."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pillow>=10.0.0",
]

[project.scripts]
generate-cards = "src.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
